Do some include untangling, to limit the spread of the Windows.h include from SevenZipFileReader.h
