+++ Core/MIPS/IR/IRNativeCommon.h
@@ -18,6 +18,7 @@
 #pragma once
 
 #include <unordered_map>
+
 #include "Core/MIPS/IR/IRJit.h"
 #include "Core/MIPS/JitCommon/JitBlockCache.h"
