+++ UI/UI.vcxproj
@@ -105,6 +105,7 @@
     <ClInclude Include="CwCheatScreen.h" />
     <ClInclude Include="GPUDriverTestScreen.h" />
     <ClInclude Include="IAPScreen.h" />
+    <ClInclude Include="ImDebugger\ImCommand.h" />
     <ClInclude Include="ImDebugger\ImConsole.h" />
     <ClInclude Include="ImDebugger\ImDebugger.h" />
     <ClInclude Include="ImDebugger\ImDisasmView.h" />
