+++ UI/UI.vcxproj.filters
@@ -291,6 +291,9 @@
     <ClInclude Include="AdhocServerScreen.h">
       <Filter>Screens</Filter>
     </ClInclude>
+    <ClInclude Include="ImDebugger\ImCommand.h">
+      <Filter>ImDebugger</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="Screens">
